requirements.txt
langchain
langchain-google-genai
langchain-mcp-adapters
langgraph
python-dotenv
google-generativeai
